+++ stashjs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Fixed
+
+- Range queries on date or uint64 would immediately throw errors
+
 ## [0.4.5]
 
 ## Changed